add auth, algolia search, lw tv chart + volume legend, visual alerts, bug fixes

+++ src/handlers/editTicker.js
@@ -3,25 +3,27 @@ import { Markup } from "telegraf";
 
 export const handleEditTickerField = async (ctx, session) => {
   const fieldData = ctx.message.text;
-  const fieldName = session.sessionData.field;
-  try {
-    await Ticker.updateField(session.sessionData.symbol, fieldName, fieldData);
-    await ctx.reply(
-      `Ticker field [${fieldName}] updated to ${fieldData}.`,
-      Markup.inlineKeyboard([
-        [
-          Markup.button.callback(
-            `Show ticker ${session.sessionData.symbol}`,
-            `show-ticker/${session.sessionData.symbol}`,
-          ),
-        ],
-      ]),
+  const { symbol, field } = session.sessionData;
+  await Ticker.updateField(symbol, field, fieldData, session);
+  await ctx.reply(
+    `Ticker field [${field}] updated to ${fieldData}.`,
+    Markup.inlineKeyboard([
+      [
+        Markup.button.callback(
+          `Show ticker ${symbol}`,
+          `show-ticker/${symbol}`,
+        ),
+      ],
+    ]),
+  );
+  if (field === "alertDown") {
+    //next step to set alert2
+    session.sessionData.field = "alertUp";
+    await session.save("bybit-scene");
+    await ctx.replyWithHTML(
+      `Editing <b>${symbol}</b>, please set <b>Alert Up</b>. Please enter a new value:`,
     );
+  } else {
     await session.delete("bybit-scene");
-  } catch (error) {
-    console.log(error);
-    ctx.reply(
-      `Error edit ticker ${session.sessionData.symbol}: ${error.message}`,
-    );
   }
 };
